refactor(e2e): auto-handle hardcoded account in setup, cleanups
- EpochsTestContext.setup auto-creates/registers hardcoded account when
  skipInitialSequencer is set (no manual handling in each test)
- Move schnorr_hardcoded_account_contract.ts from e2e_epochs/ to fixtures/
- Delete unused deployTestAccounts/registerTestAccounts methods
- Remove unnecessary setupWalletOnNode calls in P2P tests that don't
  need the wallet on a P2P node
- Remove redundant disableValidator: false from createValidatorConfig
- Move cross-chain contract deploy to test body in epochs_mbps.parallel
