+++ yarn-project/end-to-end/src/e2e_epochs/epochs_mbps.parallel.test.ts
@@ -1,4 +1,3 @@
-import type { InitialAccountData } from '@aztec/accounts/testing';
 import type { Archiver } from '@aztec/archiver';
 import type { AztecNodeService } from '@aztec/aztec-node';
 import { AztecAddress, EthAddress } from '@aztec/aztec.js/addresses';
@@ -71,19 +70,15 @@ describe('e2e_epochs/epochs_mbps', () => {
     minTxsPerBlock?: number;
     maxTxsPerBlock?: number;
     buildCheckpointIfEmpty?: boolean;
-    deployCrossChainContract?: boolean;
     skipPushProposedBlocksToArchiver?: boolean;
   }) {
-    const { syncChainTip = 'checkpointed', deployCrossChainContract = false, ...setupOpts } = opts;
+    const { syncChainTip = 'checkpointed', ...setupOpts } = opts;
 
     validators = times(NODE_COUNT, i => {
       const privateKey = bufferToHex(getPrivateKeyFromIndex(i + 3)!);
       const attester = EthAddress.fromString(privateKeyToAccount(privateKey).address);
       return { attester, withdrawer: attester, privateKey, bn254SecretKey: new SecretValue(Fr.random().toBigInt()) };
     });
-
-    // Pre-compute hardcoded account data so it gets funded in genesis.
-    const accountData: InitialAccountData = await EpochsTestContext.getHardcodedAccountData(Fr.random(), Fr.random());
 
     // Setup context with the given set of validators and MBPS configuration.
     // Timing calculation for 3 blocks per checkpoint with 8s sub-slots:
@@ -94,7 +89,6 @@ describe('e2e_epochs/epochs_mbps', () => {
     // - Total: 0.5 + 24 + 8 + 2.5 = 35s → use 36s for margin
     test = await EpochsTestContext.setup({
       numberOfAccounts: 0,
-      initialFundedAccounts: [accountData],
       initialValidators: validators,
       mockGossipSubNetwork: true,
       disableAnvilTestWatcher: true,
@@ -122,13 +116,13 @@ describe('e2e_epochs/epochs_mbps', () => {
 
     ({ context, logger, rollup } = test);
     wallet = context.wallet;
-
-    // Register the hardcoded account in PXE (local only, no on-chain deployment).
-    from = await test.registerHardcodedAccount(accountData);
+    from = context.accounts[0]; // auto-created by setup
 
     // Start the validator nodes
     logger.warn(`Initial setup complete. Starting ${NODE_COUNT} validator nodes.`);
-    nodes = await asyncMap(validators, ({ privateKey }) => test.createValidatorNode([privateKey]));
+    nodes = await asyncMap(validators, ({ privateKey }) =>
+      test.createValidatorNode([privateKey], { dontStartSequencer: true }),
+    );
     logger.warn(`Started ${NODE_COUNT} validator nodes.`, { validators: validators.map(v => v.attester.toString()) });
 
     // Point the wallet at a validator node. The initial node-0 has all validator keys in its config,
@@ -136,16 +130,6 @@ describe('e2e_epochs/epochs_mbps', () => {
     // the wallet to a validator node, the PXE correctly tracks proposed blocks.
     wallet.updateNode(nodes[0]);
     archiver = nodes[0].getBlockSource() as Archiver;
-
-    // Deploy cross-chain contract if needed (validators are running so blocks can be mined).
-    if (deployCrossChainContract) {
-      logger.warn(`Deploying cross-chain test contract`);
-      ({ contract: crossChainContract } = await TestContract.deploy(wallet).send({ from }));
-      logger.warn(`Cross-chain test contract deployed at ${crossChainContract!.address}`);
-    }
-
-    // Stop sequencers so the test body can configure and restart them.
-    await Promise.all(nodes.map(n => n.getSequencer()!.stop()));
 
     // Register contract for sending txs.
     contract = await test.registerTestContract(wallet);
@@ -280,7 +264,15 @@ describe('e2e_epochs/epochs_mbps', () => {
   });
 
   it('builds multiple blocks per slot with L2 to L1 messages', async () => {
-    await setupTest({ syncChainTip: 'proposed', minTxsPerBlock: 1, maxTxsPerBlock: 2, deployCrossChainContract: true });
+    await setupTest({ syncChainTip: 'proposed', minTxsPerBlock: 1, maxTxsPerBlock: 2 });
+
+    // Start sequencers first, then deploy cross-chain contract (needs running sequencer to mine).
+    await Promise.all(nodes.map(n => n.getSequencer()!.start()));
+    logger.warn(`Started all sequencers`);
+
+    logger.warn(`Deploying cross-chain test contract`);
+    ({ contract: crossChainContract } = await TestContract.deploy(wallet).send({ from }));
+    logger.warn(`Cross-chain test contract deployed at ${crossChainContract!.address}`);
 
     // Pre-prove all L2→L1 message transactions
     const l2ToL1Recipient = EthAddress.fromString(context.deployL1ContractsValues.l1Client.account.address);
@@ -293,10 +285,6 @@ describe('e2e_epochs/epochs_mbps', () => {
       ),
     );
     logger.warn(`Pre-proved ${txs.length} L2→L1 message transactions`);
-
-    // Start the sequencers
-    await Promise.all(nodes.map(n => n.getSequencer()!.start()));
-    logger.warn(`Started all sequencers`);
 
     // Send all transactions at once
     const txHashes = await Promise.all(txs.map(tx => tx.send({ wait: NO_WAIT })));
@@ -328,7 +316,15 @@ describe('e2e_epochs/epochs_mbps', () => {
   });
 
   it('builds multiple blocks per slot with L1 to L2 messages', async () => {
-    await setupTest({ syncChainTip: 'proposed', minTxsPerBlock: 1, maxTxsPerBlock: 1, deployCrossChainContract: true });
+    await setupTest({ syncChainTip: 'proposed', minTxsPerBlock: 1, maxTxsPerBlock: 1 });
+
+    // Start sequencers first, then deploy cross-chain contract (needs running sequencer to mine).
+    await Promise.all(nodes.map(n => n.getSequencer()!.start()));
+    logger.warn(`Started all sequencers`);
+
+    logger.warn(`Deploying cross-chain test contract`);
+    ({ contract: crossChainContract } = await TestContract.deploy(wallet).send({ from }));
+    logger.warn(`Cross-chain test contract deployed at ${crossChainContract!.address}`);
 
     const L1_TO_L2_COUNT = 4;
     const FILLER_TX_COUNT = 5; // Enough txs to advance the chain so messages become ready
@@ -350,16 +346,12 @@ describe('e2e_epochs/epochs_mbps', () => {
     });
     logger.warn(`Seeded ${l1ToL2Messages.length} L1→L2 messages`);
 
-    // Pre-prove filler txs before starting sequencers (using unique nullifiers to avoid conflicts)
+    // Pre-prove filler txs (using unique nullifiers to avoid conflicts)
     logger.warn(`Pre-proving ${FILLER_TX_COUNT} filler txs to advance the chain`);
     const fillerTxs = await timesAsync(FILLER_TX_COUNT, i =>
       proveInteraction(wallet, contract.methods.emit_nullifier(new Fr(1000 + i)), { from }),
     );
     logger.warn(`Pre-proved ${fillerTxs.length} filler txs`);
-
-    // Start the sequencers
-    await Promise.all(nodes.map(n => n.getSequencer()!.start()));
-    logger.warn(`Started all sequencers`);
 
     // Send all filler txs at once (without waiting for them to be mined)
     const fillerTxHashes = await Promise.all(fillerTxs.map(tx => tx.send({ wait: NO_WAIT })));
