+++ yarn-project/end-to-end/src/e2e_p2p/gossip_network.test.ts
@@ -147,7 +147,6 @@ describe('e2e_p2p_network', () => {
     // We need to `createNodes` before we setup account, because
     // those nodes actually form the committee, and so we cannot build
     // blocks without them (since targetCommitteeSize is set to the number of nodes)
-    t.setupWalletOnNode(nodes[0]);
     await t.setupAccount();
 
     // Wait until the other nodes sync to the block from which we sent the tx
