+++ yarn-project/end-to-end/src/e2e_p2p/rediscovery.test.ts
@@ -70,7 +70,6 @@ describe('e2e_p2p_rediscovery', () => {
     // We need to `createNodes` before we setup account, because
     // those nodes actually form the committee, and so we cannot build
     // blocks without them (since targetCommitteeSize is set to the number of nodes)
-    t.setupWalletOnNode(nodes[0]);
     await t.setupAccount();
 
     // stop bootstrap node
