+++ yarn-project/end-to-end/src/fixtures/schnorr_hardcoded_account_contract.ts
