+++ yarn-project/end-to-end/src/fixtures/setup_p2p_test.ts
@@ -206,7 +206,6 @@ export async function createValidatorConfig(
   const nodeConfig: AztecNodeConfig = {
     ...config,
     ...p2pConfig,
-    disableValidator: false,
     validatorPrivateKeys: new SecretValue(attesterPrivateKeys),
     sequencerPublisherPrivateKeys: [new SecretValue(attesterPrivateKeys[0])],
   };
